Removed a bug in the board file.
The pll locking was put in the #ifdef CONFIG_DDR_256MB_STACKED condition only.
If CONFIG_DDR_256MB_STACKED is not defined( say, a single die device), x-loader
would hang because pll was not allowed to lock.
The pll locking is now put outside the condition so that pll gets a chance to lock
irrespective of the condition.